Move AMQP stuff into its own module
